// react-vite-khoaluan/src/types/backend.d.ts
export interface IModelPaginate<T> {
  meta: {
    page: number;
    pageSize: number;
    pages: number;
    total: number;
    hasMore?: boolean;
  };
  result: T[];
}
// ...
export interface ICandidate {
  score: number;
  user: IUser; // Dùng lại interface IUser đã có
}
// ...
export interface IUser {
  id?: number | undefined;
  name?: string | undefined;
  email?: string | null;
  password?: string;
  age?: number;
  gender?: string;
  address?: string | null;
  avatar?: string | null;
  mainResume?: string | null;
  onlineResume?: IOnlineResume | null;
  workExperiences?: IWorkExperience[];
  public?: boolean;
  vipExpiryDate?: string | null | undefined;
  vip?: boolean | null;
  role?: {
    id: string;
    name: string;
  } | null;
  company?: {
    id: string;
    name: string;
  } | null;
  createdBy?: string;
  isDeleted?: boolean;
  deletedAt?: boolean | null;
  createdAt?: string;
  updatedAt?: string;
}
// ...
export type IMeta = IModelPaginate<any>["meta"];
// ...
export interface IWorkExperience {
  id?: number;
  companyName: string;
  startDate: string;
  endDate: string;
  description: string;
  location: string;
  // Thêm các trường khác nếu cần
  createdAt?: string;
  updatedAt?: string;
}

export interface IOnlineResume {
  id?: number;
  title: string;
  fullName: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  address: string;
  summary: string;
  certifications: string;
  educations: string;
  languages: string;
  skills: { id: number; name?: string }[];
  workExperiences?: IWorkExperience[]; // có thể thêm sau này
  user?: {
    id: number;
    email: string;
  };
  createdAt?: string;
  updatedAt?: string;
}